Made recipe page with three recipes
Recipes include Tiramisu, Smash Burger, and Waffle.  Each recipe contains a title, an image of the food,
a short description, a list of ingredients, and a list of steps.  The main home page has a link to all
three recipes, and each recipe page links back to the home page.